Build mixed Discover timeline
