+++ frontend/assets/scss/3-component/_articles.scss
@@ -1,36 +1,13 @@
-.articles_section {
-  margin-top: 66px;
-  padding: 32px 0;
-  background-color: var(--mm-color-secondary-light);
-}
-
-.articles_list {
-  display: flex;
-  flex-direction: row;
-  gap: 32px;
-  padding: 0 26px;
-  overflow-x: scroll;
-  scrollbar-width: none;
-  -webkit-overflow-scrolling: touch;
-  overscroll-behavior-x: contain;
-}
-
-.articles_list::-webkit-scrollbar {
-  display: none;
-}
-
 .article {
   display: flex;
-  flex: 0 0 75vw;
   flex-direction: column;
-  min-width: 75vw;
+  width: 100%;
   overflow: hidden;
   overflow: clip;
   border: var(--mm-border-default);
   border-radius: var(--mm-radius-xl);
   background-color: var(--mm-color-surface);
   box-shadow: none;
-  scroll-snap-align: start;
 }
 
 .article .content {
